Merge authors and show contribution
Sometimes authors commit with different email addresses or
change them over time. This commit changes the key in the authors
dictionary to be commit.author.name to merge such "virtual committers"
(if they have differing user names there's not much we can do about it).

In addition this commit also displays the contribution percentage
(since that has already been calculated), but only if there's more than
one contributor to the page.
Eventually I think this should be made configurable (switch on/off
and a template), but this is a more involved consideration about
the interface that I don't want to "do" without a discussion.