Optionally hide email addresses in rendered output
Sometimes, it may be prudent/desirable just to render the authors'
names, but not their email address.

In order to enable that, introduce a new configuration option,
show_email_address, which defaults to true. When true, it retains the
prior behaviour of rendering authors' names as "mailto:" links. When
false, it only renders the authors' names in plain text.

Also, add a unit test to check the new option's behaviour, and update
the documentation to explain its purpose.